Parsing(cleanup): Reorganizes tokenizing function into appropriate file #2
